LICENSE
Copyright © [Year] [Your Company/Name]. All rights reserved.

Permission is hereby granted, free of charge, to any individual or organization obtaining a copy of this software and associated documentation files (the \"Software\"), to use the Software solely for personal, non-commercial purposes, and only for evaluation.

The Software may not be copied, modified, merged, published, distributed, sublicensed, or sold, in whole or in part, without explicit written permission from [Your Company/Name].

This software is provided \"as is,\" without warranty of any kind, express or implied, including but not limited to the warranties of merchantability, fitness for a particular purpose, and noninfringement. In no event shall the authors or copyright holders be liable for any claim, damages, or other liability, whether in an action of contract, tort, or otherwise, arising from, out of, or in connection with the software or the use or other dealings in the software.